chore: move verify signature to atoma-utils
