+++ atoma-service/Cargo.toml
@@ -20,7 +20,6 @@ blake2 = { workspace = true }
 clap = { workspace = true }
 config = { workspace = true }
 dotenv = { workspace = true }
-fastcrypto = { workspace = true }
 flume = { workspace = true }
 futures = { workspace = true }
 hex = { workspace = true }
