+++ atoma-utils/Cargo.toml
@@ -6,4 +6,8 @@ license.workspace = true
 
 [dependencies]
 anyhow.workspace = true
+axum.workspace = true
+fastcrypto.workspace = true
+sui-sdk = { workspace = true }
 tokio.workspace = true
+tracing.workspace = true
